Implement receiveAttack and fix initialization of boards
This commit implements the receiveAttack and checkGameOver methods. In addition, when
testing receiveAttack, found that making an Array the way previously created a single array where
all of the elements referenced the same array. This is due to array fill populating every element with
the same reference to an array. To fix this, the lines were changed to create a new Array from an array
of the desired size with the callback function to create a new array of the second dimension's size.